mmap读操作也进行tryGrow判断

+++ bloom/options.go
@@ -2,12 +2,12 @@ package bloom
 
 type Options struct {
 	file    string
-	maxSize int
+	maxSize int64
 }
 
 type Option func(*Options)
 
-func WithMmap(file string, maxSize int) Option {
+func WithMmap(file string, maxSize int64) Option {
 	return func(o *Options) {
 		o.file = file
 		o.maxSize = maxSize
